Updated Campaign List Refresh page
This includes an updated Campaign List Refresh page which the user lands
on after clicking the "Refresh Campaign List" button on the Campaign
detail page.

This page has been updated to be a SLDS styled page that includes a
periodically updating status message to indicate to the user the update
status of their campaign, with a job progress component to also show
them periodically updated status of running apex batch jobs.

This also includes updates to the Campaign status values, whereas before
the status values were 'RUNNING', 'SUCCESS', and 'FAILURE', now there is
the expanded and normalized list of 'Queued', 'Processing', 'Aborted',
'Completed', and 'Failed'.  The status ENUM was updated, as well as all
other references to the old status enums or strings.

The updating of Campaign status and the unlocking of the Campaign has
been moved from CampaignListToCampaignBatch to
DeleteCampaignListMembersBatch.  While the Campaign is technically done
being updated after CampaignListToCampaignBatch populates the Campaign's
CampaignMembers, it was not safe to unlock the Campaign at this point.
I noticed that starting a new campaign refresh before the final cleanup
of the campaign list members was completed would cause failures.

The DeleteCampaignListMembersBatch has been updated to now be aware of
whether it is running as a final cleanup pass or not, so that it now
knows when it should update the Campaign and release the Campaign mutex
lock.  (Note that this batch class runs at both the beginning and end of
the list refresh process, so it needs to know how to differentiate
between the two executions). This includes new properties on the class,
a new constructor for setting those properties, and a finish() method
that performs the updating and unlocking.

Brought over JobProgress and PageMessages component from Cumulus with
some small modifications to PageMessages to reformat the messages using
the slds-notify-toast style (as opposed to block inline messages), and
some small modifications to JobProgress to improve visual layout, to
remove dependencies on Cumulus specific code, and to provide new tests
(again without Cumulus dependencies).  The JobProgress component uses
the moment.js and transparency.js libraries, so these have been included
as static resources.