Add error trapping to Campaign Tools batches
Add CampaignList.Service.unlockCampaignAndMarkFailed() method.  This is
a convenience method intended to be called from the error handling code
in the various batch classes.  It will unlock the campaign, mark the
status as 'Failed', and also log any provided error message to the
specified campaign.

Refactor stub test classes to make naming convention more uniform.
Remove some special case code from stub classes to stub sub classes
where some tests needed special behavior (like logging a sequence of
calls, as opposed to just the first call, etc.).  All test classes have
been updated to use the new naming convention.

Add some system class wrappers to service class.  This gives the ability
to stub calls to BatchableContext.getJobId() and System.abortJob() to
verify classes are correctly calling those methods, but not need to
create real jobs in order to get test coverage.

Since batch jobs now need to be able to unlock the campaign that is
being refreshed, all batches now need to be aware of the target campaign
of the refresh process.  All batch classes have been updated to include
a reference to this campaign by id, and this has been added to the
constructor signatures where necessary.

Each batch class has had a try/catch block added to wrap the contents of
the execute() method, and in some cases also the start() method.  This
includes: CampaignListFromCampaignBatch, CampaignListFromReportBatch,
CampaignListToCampaignBatch, and DeleteCampaignListMembersBatch.
DeleteCampaignMembersBatch was not modified, since it is expected to be
removed as part of future work.

New tests to verify exception handling in batch classes.

CampaignListFromReportBatch.IdIterablyFactory interface was added to
allow stubbing the creation of the iterable in the start() method.  This
was necessary in order to be able to trigger a fake exception from a
test in order to verify the exception handling of the start() method.

New labels added for error messages that are created as part of the
exception handling in batch classes.

New field added to Campaign for storing error message from batch.