Modify how campaigns are refreshed
Previously, all existing CampaignMember records were deleted from a
Campaign before refreshing it using a Campaign List.  This was
undesireable, because it would unnecessarily delete information that
might be stored on CampaignMember records where those CampaignMember
records would ultimately be re-populated onto the Campaign.  This work
alters the way a Campaign is refreshed to avoid deleting CampaignMembers
that would be retained after refresh.

DeleteCampaignMembersBatch was removed, since we no longer perform this
step at the beginning of the refresh process.

The CampaignListToCampaignBatch was removed and replaced with two other
batches that refresh the Campaign in a two step process.

These refresh batch classes delegate to methods in
CampaignListMemberMapper, so this class was modified to remove the
addMembersToCampaign() method (previously used by
DeleteCampaignMembersBatch) and to add the insertCampaignMembers()
method (used by CampaignListInsertMembersBatch) and the
updateAndDeleteCampaignMembers() method (used by
CampaignListUpdateMembersBatch).

CampaignListUpdateMembersBatch was added as step one in the populating
Campaign stage of the refresh process.  This batch will process all
existing CampaignMember records and, for those CampaignMember records
that have existing temporary Member__c records, will update the
CampaignMember records where the corresponding Member__c record meets
the campaign list criteria, otherwise the existing CampaignMember record
will be deleted (as it no longer meets the campaign list criteria).

CampaignListInsertMembersBatch was added as step two in the populating
Campaign stage of the refresh process.  This batch will process all
temporary Member__c records and, for those Member__c records that meet
the given campaign list criteria and that do not have an existing
CampaignMember record already created, will insert a new CampaignMember
record for that temporary Member__c record.

CampaignListService.updateCampaignFromCampaignList() method was updated
to remove DeleteCampaignMembersBatch as the first step in the refresh
process.  The CampaignListToCampaignBatch step was replaced with the
CampaignListUpdateMembersBatch and CampaignListInsertMembersBatch steps.

A MemberStub class was added to CampaignList_TEST to support mocking
objects that represent temporary Member__c records in tests.  Like all
other stub classes, it simply stores the arguments passed to methods and
returns pre-specified return values from methods.