Restructure of Campaign List editor LC
Removed CampaignListApp, CampaignListCmp, Source, SourceMultiList, addCSegmentEvent, and deleteCSegmentEvent Aura component bundles.

Replaced with CampaignListEditorApp, CampaignListEditor, CampaignListSource, CampaignListSegment, and CampaignListSegmentUpdateEvent Aura bundles.

The major change here was moving responsibility for updating the campaign list tree from within each sub-component to the top level editor component.  Subcomponents no longer modify their own segments, but they emit messages that the top level component listens to and makes modifications to the segment tree.  The motivation for this change was to workaround strange/buggy behavior observed when using <aura:iterate> on arrays that were being structurally modified by components other than the one in which the <aura:iterate> component was used.

Also, additional SLDS styling was applied to the editor components.  All strings in the editor lightning components and CampaignList page now use custom labels.  One unfortunate result of using custom labels in lightning components is that a page, LightningComponentLabels, needed to be created to reference those labels, because otherwise the package spider will not detect those labels for inclusion in this package.  This is considered the "recommended" approach according to the Aura documentation :-/

A new PageMessages lightning component was created to allow the display of SLDS styled toast messages from withing a Lightning Component.  This is intended to be an SLDS styled replacement for <ui:message>.  This component allows for displaying of messages by other components or from the hosting visualforce page by listening for an application level event, AddPageMessageEvent.

The CampaignList page was modified to include an SLDS style page header, and to allow display of an Aloha header for users that are not in LEX.  The 'Save' button no longer exists within the Editor component itself, but instead the editor component listens to a CampaignEditorSaveEvent which is emitted by the Save button on the VF page.

A bug in CampaignListSegmentMapper that would cause empty source segments to get saved to the database, or fail deserialization altogether, was resolved.  Now if a source element is being saved from the campaign list editor, it must have a source id specified.
The button on the Campaign detail page that said "Edit Campaign List" now says "Define Campaign List".